// enums.ts
export enum ServicesEnum {
  POSTGRE = 'postgresql@12',
  MYSQL = 'mysql@5.7',
}

export enum ScriptsEnum {
  POSTGRE_START = 'POSTGRE_START',
  POSTGRE_RESTART = 'POSTGRE_RESTART',
  POSTGRE_STOP = 'POSTGRE_STOP',
  POSTGRE_CONNECT = 'POSTGRE_CONNECT',
  MYSQL_START = 'MYSQL_START',
  MYSQL_RESTART = 'MYSQL_RESTART',
  MYSQL_STOP = 'MYSQL_STOP',
  MYSQL_CONNECT = 'MYSQL_CONNECT',
  GET_RANDOM_HEX = 'GET_RANDOM_HEX',
  GET_RANDOM_NUMBER = 'GET_RANDOM_NUMBER',
  ZSHRC_EDIT = 'ZSHRC_EDIT',
  HOSTS_EDIT = 'HOSTS_EDIT',
  PROCESS_ID_BY_PORT_GET = 'PROCESS_ID_BY_PORT_GET',
  PROCESS_BY_PID_KILL = 'PROCESS_BY_PID_KILL',
}

export enum VariablesEnum {
  LOCAL_PORT = 'LOCAL_PORT',
  PROCESS_ID = 'PROCESS_ID',
  NUMBER_FROM = 'NUMBER_FROM',
  NUMBER_TO = 'NUMBER_TO',
  HEX_LENGTH = 'HEX_LENGTH',
}

export enum VariablesTypesEnum {
  NUMBER = 'number',
  STRING = 'text',
}